requirements-CI.txt
pytest==6.2.4
apache-airflow-backport-providers-google==2021.3.3
apache-airflow-backport-providers-cncf-kubernetes